CodeEngineVB/requirements.txt
ai-core-sdk
flask==2.2
python-dotenv